Extract context generation to a separate component
